removed self discovery add robust node verification for routes created new files for node peer and node routes to remove circular import
